project/backend/app/schemas/token.py
 """
 JWT token schemas.
 """

from typing import Optional

 from pydantic import BaseModel


 class Token(BaseModel):
     """Access token response."""

     access_token: str
     token_type: str = "bearer"


 class TokenPayload(BaseModel):
     """Payload decoded from JWT."""

     sub: str
     exp: Optional[int] = None
     iat: Optional[int] = None
     nbf: Optional[int] = None
