project/backend/pyproject.toml
 [tool.poetry]
 name = "healthcare-platform-backend"
 version = "0.1.0"
 description = "FastAPI backend for the healthcare appointment platform"
 authors = ["AI Assistant <assistant@example.com>"]
 readme = "README.md"
 package-mode = false

 [tool.poetry.dependencies]
 python = "^3.11"
 fastapi = "^0.115.0"
 uvicorn = { version = "^0.30.0", extras = ["standard"] }
 sqlalchemy = "^2.0.30"
 alembic = "^1.13.1"
 psycopg = { version = "^3.1.18", extras = ["binary"] }
 python-jose = "^3.3.0"
 passlib = { version = "^1.7.4", extras = ["bcrypt"] }
 pydantic-settings = "^2.3.1"
 redis = "^5.0.4"
 httpx = "^0.27.0"
 email-validator = "^2.2.0"

 [tool.poetry.group.dev.dependencies]
 pytest = "^8.3.0"
 pytest-asyncio = "^0.23.6"
 coverage = "^7.5.0"
 ruff = "^0.4.10"
 mypy = "^1.10.0"

 [build-system]
 requires = ["poetry-core>=1.7.0"]
 build-backend = "poetry.core.masonry.api"
